added Firefox test case
I added the firefox test cases. Additionally, I attempted to create additional helper functions, but I was running into issues regarding how to parse the string. It would clean up the code a bit, but it does not look too bad the way it is currently using the "IF" statements over and over.