must figure out bash
This case is correct...I can visually inspect that we are passing this test case; however, I am having trouble ensuring that it is correct inside of bash. Bash works a bit weirdly with strings and I currently am having a bit of trouble working with them. I will revisit this later.